+++ workspaces/scaffolder-backend-module-kubernetes/plugins/kubernetes-actions/src/module.ts
@@ -18,7 +18,7 @@ import {
   createBackendModule,
 } from '@backstage/backend-plugin-api';
 import { CatalogClient } from '@backstage/catalog-client';
-import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
+import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node';
 
 import { createKubernetesNamespaceAction } from './actions';