+++ workspaces/linguist/plugins/linguist-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-linguist-common
 
+## 0.13.0
+
+### Minor Changes
+
+- 1e80a16: Backstage version bump to v1.44.0
+
 ## 0.12.0
 
 ### Minor Changes