+++ workspaces/scaffolder-backend-module-kubernetes/plugins/kubernetes-actions/CHANGELOG.md
@@ -1,5 +1,17 @@
 ### Dependencies
 
+## 2.12.0
+
+### Minor Changes
+
+- 126f782: Backstage version bump to v1.44.1
+
+## 2.11.0
+
+### Minor Changes
+
+- c9a5d76: Backstage version bump to v1.43.3
+
 ## 2.10.1
 
 ### Patch Changes